timely progress update - project structure reorganization
