+++ src/App.jsx
@@ -2,8 +2,8 @@ import React from "react";
 import { Route, Switch } from "react-router-dom";
 import Login from "./layouts/Login.jsx";
 import Main from "./layouts/Main.jsx";
-import UsersList from "./components/UsersList.jsx";
-import NavBar from "./components/NavBar.jsx";
+import Users from "./layouts/Users.jsx";
+import NavBar from "./components/ui/NavBar.jsx";
 import NotFound from "./components/NotFound.jsx";
 
 export const App = () => {
@@ -13,7 +13,7 @@ export const App = () => {
             <Switch>
                 <Route exact path="/" component={Main}/>
                 <Route path="/login" component={Login} />
-                <Route path="/users/:userId?" component={UsersList} />
+                <Route path="/users/:userId?" component={Users} />
                 <Route component={NotFound} />
             </Switch>
         </>
