+++ src/components/common/page/userListPage/index.js
@@ -1,0 +1,1 @@
+export { UsersListPage } from "./UsersListPage.jsx";
