+++ src/components/common/page/userPage/index.js
@@ -1,0 +1,1 @@
+export { UserPage } from "./UserPage.jsx";
