+++ src/components/common/table/index.js
@@ -1,0 +1,3 @@
+export { Table } from "./Table.jsx";
+export { TableBody } from "./TableBody.jsx";
+export { TableHeader } from "./TableHeader.jsx";
