+++ src/components/ui/SearchStatus.jsx
