+++ src/components/ui/qualities/Quality.jsx
