+++ src/components/ui/qualities/index.js
@@ -1,0 +1,1 @@
+export { default as Qualities } from "./QualitiesList.jsx";
